Intent: Consolidate CLI test coherence Operators involved: [Coherence, Resonance] Affected invariants: [#8, #9] Key changes: - Merge CLI sanity/history coverage into focused suites without redundant mocks - Exercise metrics/run/sequence/version flows via real execution and history persistence - Preserve argument helper and serialization checks alongside CLI flow tests Expected risks/dissonances: Slightly longer CLI runs mitigated by minimal graph sizes Metrics: Tg_global, latency_mean sampled via CLI runs; no regression expected for C(t), Si, νf, phase Equivalence map: None
